task6/django_project/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.3
crispy-bootstrap4==2024.10
Django==4.2.20
django-bootstrap4==25.1
django-crispy-forms==2.3
pillow==11.1.0
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.12.2

djangorestframework~=3.15.2
django-filter~=25.1
djangorestframework-simplejwt==5.5.0
